requirements.txt
playwright
python-dotenv
SQLAlchemy
psycopg2
asyncpg
requests
flake8
rich
aiohttp